refactor(foldkit): move resource provision out of the boot closure
`startWith` set up a one-time build of the app's `resources` Layer
into the runtime scope, made a Ref for each Managed Resource, and
defined the two functions that hand those to Effects:
`provideAllResources` for Commands and Subscriptions, and
`provideResources` for Flags.

`makeResourceProvider` in `resourceProvider.ts` now does that setup and
returns the two provide functions and the Managed Resource refs. It
takes the `resources` Layer, the Managed Resource configs, the runtime
scope, and the port channels. The boot closure calls it right after
the port connector binds and destructures the three values, so every
use in `runtime.ts` is unchanged. The moved code is unchanged too,
apart from the type arguments on `ManagedResourceRef`, and its comments
moved with it. `ManagedResourceRef` is now a named type exported from
the new module instead of an alias derived from the refs array.

One thing moved in time: the cached Layer build and the Refs are now
created after the port connector binds, where before the cached build
came first and the Refs after. Both are plain allocations. Nothing is
built or registered in the scope until the first Command, Subscription,
or Flags run asks for a resource, so the order of teardown is the same.

No user-facing effect, so the changeset is empty.
